+++ CHANGELOG.md
@@ -11,7 +11,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Use explicit allocator for QuantityArray?
 * Add Python wrapper for obj & ply w/ interop to other classes
 * Support for numpy bytes/unitcode scalars arrays
-* Add additional tests for shared properties w/ errors for missing required properties and normalization of additional types within allOf (null, int, uint, int64, uint64, double, yggdrasil object, yggdrasil string
+* Add tests for normalization of additional types within allOf (yggdrasil object, yggdrasil string
+* Add test for conflicting aliases in allOf set
+* Add test for failed normalization of default in shared property
+* Normalize default outside of shared property
+* Add test for array of strings & complex
+* Add test for failed subtype on base scalar (e.g. double, int)
+* Add tests for missing schema properties (subtype, precision, shape), including on base scalar
+* Add test for validating ndarray without shape specified
 
 ### IN PROGRESS
 * Add docstrings to new methods